requirements.txt
Flask==2.3.2
requests==2.30.0
xmltodict==0.13.0
pymongo==4.3.3
python-dotenv==1.0.0